+++ test/plugin/rack_file_test.rb
@@ -6,12 +6,13 @@ describe "the rack_file plugin" do
   end
 
   it "enables assignment of Rack file hashes" do
-    uploaded_file = @attacher.assign({
+    @attacher.assign({
       tempfile: fakeio("image"),
       filename: "image.jpg",
       type: "image/jpeg",
       head: "...",
     })
+    uploaded_file = @attacher.get
     assert_equal "image",      uploaded_file.read
     assert_equal 5,            uploaded_file.size
     assert_equal "image.jpg",  uploaded_file.original_filename
@@ -20,13 +21,15 @@ describe "the rack_file plugin" do
 
   it "adds #path, #to_io and #tempfile methods to IO" do
     @attacher.cache.instance_eval do
-      def process(io, context)
-        io.path
-        io.to_io
-        io.tempfile
+      def upload(io, context = {})
+        @rack_file = io
+        super
       end
     end
-
-    @attacher.assign({tempfile: Tempfile.new("")})
+    @attacher.assign({tempfile: tempfile = Tempfile.new("")})
+    rack_file = @attacher.cache.instance_variable_get("@rack_file")
+    refute_empty rack_file.path
+    assert_equal tempfile, rack_file.to_io
+    assert_equal tempfile, rack_file.tempfile
   end
 end